# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API untuk platform validasi hasil kerja AI oleh ahli manusia.",
        "title": "AIValid API",
        "termsOfService": "https://aivalid.id/rules-content",
        "contact": {
            "name": "AIValid Support",
            "url": "https://aivalid.id/contact-support"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "1.0"
    },
    "host": "api.aivalid.id",
    "basePath": "/api/v1",
    "paths": {
...
        "/validation-cases/{id}/consultation-requests": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "List all consultation requests for a validation case. Only the case owner can view.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ValidationCases"
                ],
                "summary": "List consultation requests",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Validation Case ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerConsultationRequestListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Submit a consultation request for a validation case. Requires authentication.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ValidationCases"
                ],
                "summary": "Submit consultation request",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Validation Case ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerIDResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/validation-cases/{id}/consultation-requests/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get the current user's consultation request for a validation case.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ValidationCases"
                ],
                "summary": "Get my consultation request",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Validation Case ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerConsultationRequestSingleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/validation-cases/{id}/consultation-requests/{requestId}/approve": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Approve a consultation request. Only the case owner can approve.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ValidationCases"
                ],
                "summary": "Approve consultation request",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Validation Case ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Consultation Request ID",
                        "name": "requestId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerStatusResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/validation-cases/{id}/consultation-requests/{requestId}/reject": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Reject a consultation request with a reason. Only the case owner can reject.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ValidationCases"
                ],
                "summary": "Reject consultation request",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Validation Case ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Consultation Request ID",
                        "name": "requestId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Rejection reason",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerRejectConsultationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerStatusResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/validation-cases/{id}/contact": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Reveal the case owner's Telegram contact. Only approved validators can access.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ValidationCases"
                ],
                "summary": "Get case contact info",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Validation Case ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerContactRevealResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.SwaggerConsultationRequestListResponse": {
            "type": "object",
            "properties": {
                "consultation_requests": {
                    "type": "array",
                    "items": {}
                }
            }
        },
        "handlers.SwaggerConsultationRequestSingleResponse": {
            "type": "object",
            "properties": {
                "consultation_request": {}
            }
        },
        "handlers.SwaggerContactRevealResponse": {
            "type": "object",
            "properties": {
                "telegram": {
                    "type": "string",
                    "example": "@username"
                }
            }
        },
...
        "handlers.SwaggerErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "AUTH001"
                },
                "details": {
                    "type": "string",
                    "example": "Detail tambahan"
                },
                "message": {
                    "type": "string",
                    "example": "Email atau password salah"
                }
            }
        },
...
        "handlers.SwaggerIDResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 42
                }
            }
        },
...
        "handlers.SwaggerRejectConsultationRequest": {
            "type": "object",
            "properties": {
                "reason": {
                    "type": "string",
                    "example": "Alasan penolakan yang valid"
                }
            }
        },
...
        "handlers.SwaggerStatusResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AdminAuth": {
            "description": "Admin JWT token. Type \"Bearer\" followed by a space and admin JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}